pyproject.toml
[project]
name = "poll-agents"
version = "0.1.0"
description = "Poll Agents - A WebSocket server for AI agents to answer survey questions"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "aiosmtplib>=3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "aiofiles>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
supabase = [
    "supabase>=2.0",
]

[project.scripts]
poll-agents = "poll_agents.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
